+++ Services/enrollment_service.py
@@ -31,25 +31,32 @@ class EnrollmentService:
             return self.enrollment_repository.get_enrollments_by_datum_upisa(datum_upisa)
         return self.enrollment_repository.get_enrollments()
 
-    def create_enrollment(self,enrollment : schemas.EnrolmentCreate):
-        if self.enrollment_repository.get_enrollment(enrollment.student_id,enrollment.sifra_predmeta,enrollment.datum_upisa):
-            raise HTTPException(status_code=400,detail = "Enrollment already exists")
-        return self.enrollment_repository.create_enrollment(enrollment)
+    def create_enrollment(self,enrollment : schemas.EnrolmentCreate,username : str):
+        if self.enrollment_repository.is_admin(username) or self.enrollment_repository.is_student(username,enrollment.student_id) :
+            if self.enrollment_repository.get_enrollment(enrollment.student_id,enrollment.sifra_predmeta,enrollment.datum_upisa):
+                raise HTTPException(status_code=400,detail = "Enrollment already exists")
+            return self.enrollment_repository.create_enrollment(enrollment)
+        else:
+            raise HTTPException(status_code=401,detail="You are not authorized to create enrollment")
 
 
-    def update_enrollment(self ,student_id, sifra_predmeta, datum_upisa, up_enrollment):
-        enrollment  = self.enrollment_repository.get_enrollment(student_id,sifra_predmeta,datum_upisa)
-        if enrollment is None:
-            raise HTTPException(status_code=400,detail = "Enrollment doesn't exists")
-        return self.enrollment_repository.update_enrollment(enrollment,up_enrollment)
+    def update_enrollment(self ,student_id, sifra_predmeta, datum_upisa, up_enrollment, username : str):
+        if self.enrollment_repository.is_admin(username):
+            enrollment  = self.enrollment_repository.get_enrollment(student_id,sifra_predmeta,datum_upisa)
+            if enrollment is None:
+                raise HTTPException(status_code=400,detail = "Enrollment doesn't exists")
+            return self.enrollment_repository.update_enrollment(enrollment,up_enrollment)
+        else:
+            raise HTTPException(status_code=401,detail="You are not authorized to update enrollment")
 
-
-    def delete_enrollment(self, student_id, sifra_predmeta, datum_upisa):
-        enrollment = self.enrollment_repository.get_enrollment( student_id, sifra_predmeta, datum_upisa)
-        if enrollment is None:
-            raise HTTPException(status_code=400, detail="Enrollment doesn't exists")
-        return self.enrollment_repository.delete_enrollment( enrollment)
-
+    def delete_enrollment(self, student_id, sifra_predmeta, datum_upisa,username : str):
+        if self.enrollment_repository.is_admin(username):
+            enrollment = self.enrollment_repository.get_enrollment( student_id, sifra_predmeta, datum_upisa)
+            if enrollment is None:
+                raise HTTPException(status_code=400, detail="Enrollment doesn't exists")
+            return self.enrollment_repository.delete_enrollment( enrollment)
+        else:
+            raise HTTPException(status_code=401,detail="You are not authorized to delete enrollment")
 
     def get_enrollment(self, student_id, sifra_predmeta, datum_upisa):
         enrollment = self.enrollment_repository.get_enrollment( student_id, sifra_predmeta, datum_upisa)
