+++ Services/exam_services.py
@@ -1,7 +1,5 @@
 from datetime import datetime
-
 from fastapi import HTTPException,Depends
-from sqlalchemy.orm import Session
 
 from ..Schemas import schemas
 from ..Repository.exam_repository import ExamRepository,get_exam_repository
@@ -20,21 +18,30 @@ class ExamServices:
         return exam
 
 
-    def create(self,  exam: schemas.ExamCreate):
-        if self.exam_repository.get_exam( exam.student_id,exam.sifra_predmeta,exam.datum):
-            raise HTTPException(status_code=400,detail="Course already exists")
-        return self.exam_repository.create_exam( exam)
+    def create(self,  exam: schemas.ExamCreate, username : str):
+        if self.exam_repository.is_admin(username) or self.exam_repository.is_professor(username):
+            if self.exam_repository.get_exam( exam.student_id,exam.sifra_predmeta,exam.datum):
+                raise HTTPException(status_code=400,detail="Course already exists")
+            return self.exam_repository.create_exam( exam)
+        else:
+            raise HTTPException(status_code=401,detail="You are not authorized to create exam")
 
-    def delete(self, student_id:int , sifra_predmeta: str ,datum : datetime):
-        response = self.exam_repository.delete_exam( student_id,sifra_predmeta,datum)
-        if response is None:
-            raise HTTPException(status_code=404,detail="Course not found")
-        return response
+    def delete(self, student_id:int , sifra_predmeta: str ,datum : datetime, username : str):
+        if self.exam_repository.is_admin(username) or self.exam_repository.is_professor(username):
+            response = self.exam_repository.delete_exam( student_id,sifra_predmeta,datum)
+            if response is None:
+                raise HTTPException(status_code=404,detail="Course not found")
+            return response
+        else:
+            raise HTTPException(status_code=401,detail="You are not authorized to delete exam")
 
 
-    def update(self,  exam : schemas.Exam):
-        exam = self.exam_repository.update_exam( exam)
-        if exam is None:
-            raise HTTPException(status_code=404,detail="Exam not found")
-        return exam
+    def update(self,  exam : schemas.Exam, username : str):
+        if self.exam_repository.is_admin(username) or self.exam_repository.is_professor(username):
+            exam = self.exam_repository.update_exam( exam)
+            if exam is None:
+                raise HTTPException(status_code=404,detail="Exam not found")
+            return exam
+        else:
+            raise HTTPException(status_code=401,detail="You are not authorized to update exam")
