+++ Tests/test_sql_exams.py
@@ -4,7 +4,8 @@ from sqlalchemy.orm import sessionmaker
 from sqlalchemy.pool import StaticPool
 
 from ..Database.database import Base
-from ..studentski_servis import app, get_db
+from ..studentski_servis import app
+from ..dependencies import get_db
 
 SQLALCHEMY_DATABASE_URL = "sqlite:///./sql_test_app.db"
 
@@ -31,16 +32,21 @@ app.dependency_overrides[get_db] = override_get_db
 client = TestClient(app)
 
 def test_create_exam():
-    response = client.post(
-        "/exams/",
-        json={
-                  "datum": "2024-09-18T13:02:21.702Z",
-                  "ocena": 6,
-                  "polozen": True,
-                  "student_id": 1,
-                  "sifra_predmeta": "P120"
-        } ,
-    )
+
+    auth = client.post("/login/user",data={"grant_type" : "password","username": "admin", "password": "password"} )
+    assert auth.status_code == 200
+    access_token = auth.json().get("access_token")
+
+    response = client.post("/exams/",
+                            json={
+                                      "datum": "2024-09-18T13:02:21.702Z",
+                                      "ocena": 6,
+                                      "polozen": True,
+                                      "student_id": 1,
+                                      "sifra_predmeta": "P120"
+                            } ,
+                            headers= {"Authorization": f"Bearer {access_token}"}
+                            )
     assert response.status_code == 200, response.text
     data = response.json()
     assert data["datum"] == "2024-09-18T13:02:21.702000"
@@ -49,7 +55,7 @@ def test_create_exam():
     assert data["sifra_predmeta"] == "P120"
 
 
-def test_get_courses():
+def test_get_exams():
     response = client.get("/exams/1/P120/2024-09-18T13:02:21.702Z")
     assert response.status_code == 200, response.text
     data = response.json()
@@ -58,19 +64,31 @@ def test_get_courses():
     assert data["student_id"] == 1
     assert data["sifra_predmeta"] == "P120"
 
-def test_update_courses():
-    response = client.put("/exams/1/P120/2024-09-18T13:02:21.702Z", json= {
+def test_update_exams():
+    auth = client.post("/login/user", data={"grant_type": "password", "username": "admin", "password": "password"})
+    assert auth.status_code == 200
+    access_token = auth.json().get("access_token")
+
+    response = client.put("/exams/1/P120/2024-09-18T13:02:21.702Z",
+                          json= {
                               "datum": "2024-09-18T13:04:24.978Z",
                               "ocena": 9,
                               "polozen": True
-                            })
+                          },
+                          headers= {"Authorization": f"Bearer {access_token}"}
+                          )
+
     assert response.status_code == 200, response.text
     data = response.json()
     assert data["ocena"] == 9
     assert data["polozen"] == True
 
-def test_delete_student():
-    response = client.delete("/exams/1/P120/2024-09-18T13:02:21.702Z")
+def test_delete_exams():
+    auth = client.post("/login/user", data={"grant_type": "password", "username": "admin", "password": "password"})
+    assert auth.status_code == 200
+    access_token = auth.json().get("access_token")
+
+    response = client.delete("/exams/1/P120/2024-09-18T13:02:21.702Z", headers= {"Authorization": f"Bearer {access_token}"})
     assert response.status_code == 200, response.text
     data = response.json()
     assert data == {"Exam deleted" : True}
