+++ Tests/test_sql_statistics.py
@@ -4,7 +4,8 @@ from sqlalchemy.orm import sessionmaker
 from sqlalchemy.pool import StaticPool
 
 from ..Database.database import Base
-from ..studentski_servis import app, get_db
+from ..studentski_servis import app
+from ..dependencies import get_db
 
 SQLALCHEMY_DATABASE_URL = "sqlite:///./sql_test_app.db"
 
@@ -31,12 +32,15 @@ app.dependency_overrides[get_db] = override_get_db
 client = TestClient(app)
 
 def test_statistics():
+    auth = client.post("/login/user",data={"grant_type" : "password","username": "admin", "password": "password"} )
+    assert auth.status_code == 200
+    access_token = auth.json().get("access_token")
     response = client.post("/students/",
                     json={
                         "ime": "Petar",
                         "prezime": "Petrovic",
                         "indeks": "100-2023"
-                    },
+                    },headers= {"Authorization": f"Bearer {access_token}"}
     )
     data = response.json()
     assert "id" in data
@@ -48,7 +52,8 @@ def test_statistics():
                         "espb": 6,
                         "sifra_predmeta": "P120",
                         "profesor_id" : 1
-                })
+                },
+                headers= {"Authorization": f"Bearer {access_token}"})
 
     client.post("/courses/",
                 json={
@@ -56,7 +61,9 @@ def test_statistics():
                     "espb": 6,
                     "sifra_predmeta": "P121",
                     "profesor_id": 1
-    })
+                },
+                headers= {"Authorization": f"Bearer {access_token}"}
+                )
 
     client.post("/courses/",
                 json={
@@ -64,7 +71,8 @@ def test_statistics():
                     "espb": 6,
                     "sifra_predmeta": "M101",
                     "profesor_id": 1
-                })
+                },
+                headers= {"Authorization": f"Bearer {access_token}"})
 
     client.post("/courses/",
                 json={
@@ -72,49 +80,59 @@ def test_statistics():
                     "espb": 6,
                     "sifra_predmeta": "M102",
                     "profesor_id": 1
-                })
-    client.post("/exams/" , json={
+                },
+                headers= {"Authorization": f"Bearer {access_token}"})
+    client.post("/exams/" ,
+                json={
                     "datum": "2024-09-19T07:38:46.871Z",
                     "ocena": 7,
                     "polozen": True,
                     "student_id": student_id,
                     "sifra_predmeta": "P120"
-        })
-    client.post("/exams/" , json={
+                },
+                headers= {"Authorization": f"Bearer {access_token}"})
+    client.post("/exams/" ,
+                json={
                     "datum": "2024-09-19T07:38:46.871Z",
                     "ocena": 5,
                     "polozen": False,
                     "student_id": student_id,
                     "sifra_predmeta": "P121"
-        })
-    client.post("/exams/" , json={
+                },
+                headers= {"Authorization": f"Bearer {access_token}"})
+    client.post("/exams/" ,
+                json={
                     "datum": "2024-09-19T07:38:46.871Z",
                     "ocena": 8,
                     "polozen": True,
                     "student_id": student_id,
                     "sifra_predmeta": "M101"
-        })
-    client.post("/exams/" , json={
+                },
+                headers= {"Authorization": f"Bearer {access_token}"})
+    client.post("/exams/" ,
+                json={
                     "datum": "2024-09-19T07:38:46.871Z",
                     "ocena": 5,
                     "polozen": False,
                     "student_id": student_id,
                     "sifra_predmeta": "M102"
-        })
+                },
+                headers= {"Authorization": f"Bearer {access_token}"})
 
-    respone = client.get(f"/statistics/{student_id}/")
+    respone = client.get(f"/statistics/{student_id}/",headers= {"Authorization": f"Bearer {access_token}"})
     assert respone.status_code == 200
     assert respone.json() == {"ESPB" : 12, "Polozio" :2}
 
 
-    client.delete(f"/exams/{student_id}/P120/2024-09-19T07:38:46.871Z")
-    client.delete(f"/exams/{student_id}/P121/2024-09-19T07:38:46.871Z")
-    client.delete(f"/exams/{student_id}/M101/2024-09-19T07:38:46.871Z")
-    client.delete(f"/exams/{student_id}/M102/2024-09-19T07:38:46.871Z")
+    client.delete(f"/exams/{student_id}/P120/2024-09-19T07:38:46.871Z",headers= {"Authorization": f"Bearer {access_token}"})
+    client.delete(f"/exams/{student_id}/P121/2024-09-19T07:38:46.871Z",headers= {"Authorization": f"Bearer {access_token}"})
+    client.delete(f"/exams/{student_id}/M101/2024-09-19T07:38:46.871Z",headers= {"Authorization": f"Bearer {access_token}"})
+    client.delete(f"/exams/{student_id}/M102/2024-09-19T07:38:46.871Z",headers= {"Authorization": f"Bearer {access_token}"})
 
-    client.delete(f"/students/{student_id}")
-    client.delete("/courses/P120")
-    client.delete("/courses/P121")
-    client.delete("/courses/M101")
-    client.delete("/courses/M102")
+    client.delete(f"/students/{student_id}",headers= {"Authorization": f"Bearer {access_token}"})
+    client.delete("/courses/P120",headers= {"Authorization": f"Bearer {access_token}"})
+    client.delete("/courses/P121",headers= {"Authorization": f"Bearer {access_token}"})
+    client.delete("/courses/M101",headers= {"Authorization": f"Bearer {access_token}"})
+    client.delete("/courses/M102",headers= {"Authorization": f"Bearer {access_token}"})
+    client.delete("/students/2",headers= {"Authorization": f"Bearer {access_token}"})
